% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugBank.R
\name{addDBLayer}
\alias{addDBLayer}
\title{Add a drug layer to a mully graph}
\usage{
addDBLayer(g, data, drugList)
}
\arguments{
\item{g}{The mully graph}

\item{data}{The dataframe containing the parsed information of DrugBank. This argument can be obtained using the function loadDBXML(DrugBankFile)}

\item{drugList}{The list of DrugBank Ids of the drugs to be added. This argument can be either a string (one drug) or a list of strings (multiple drugs)}
}
\value{
A mully graph with the added drug layer
}
\description{
Add a drug layer to a mully graph
}
\examples{
\dontrun{ 
data=readDBXML(DBXMLFilePath)
g=mully("DrugBank",direct=T)
g=addDBLayer(g,data,c("DB00001","DB06605"))
}
}
