% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMIM.R
\name{addDiseaseLayer}
\alias{addDiseaseLayer}
\title{Add OMIM Layer with its respective Nodes}
\usage{
addDiseaseLayer(g, biopax, addGenesLayer = FALSE)
}
\arguments{
\item{g}{A Mully graph}

\item{biopax}{A Biopax object}
}
\value{
g A Mully graph including OMIM nodes and possible edges
}
\description{
Add OMIM Layer with its respective Nodes
}
\note{
should be preceded by calling  romim::set_key('KEY'). The KEY could be requested via omim's official website.
}
\examples{
\dontrun{ 
biopax=readBiopax("wnt.owl") #This is a mully graph that contains a protein layer
pathwayID=listPathways(biopax)$id[1]
g=Multipath::pathway2Mully(biopax,pathwayID)
g=addGenesLayer(g,biopax)
g=addDiseaseLayer(g,biopax)
}
}
\author{
Mohammad Al Maaz
}
