% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG.R
\name{addGenesLayer}
\alias{addGenesLayer}
\title{Add KEGG Gene Layer with its respective Nodes}
\usage{
addGenesLayer(g, geneList, biopax)
}
\arguments{
\item{g}{An existing Mully graph}

\item{geneList}{A list of KEGG genes to add}
}
\value{
A Mully graph with the all layers,nodes and edges
}
\description{
Add KEGG Gene Layer with its respective Nodes
}
\examples{
g= mully("Test")   #This is a new graph but it is allowed to have an existing graph with nodes and edges already added
geneList=c("hsa:122706","hsa:4221","hsa:8312")
up = UniProt.ws()
graph=addGenesLayer(g,geneList)
}
