% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Views.R
\name{addStep}
\alias{addStep}
\title{Track a modification of a graph}
\usage{
addStep(
  v,
  action,
  element,
  name = NA,
  layername = NA,
  V1 = NA,
  V2 = NA,
  attributes = NA,
  multi = F,
  trans = T
)
}
\arguments{
\item{v}{The input view in which the modification should be saved}

\item{action}{The type of action to be applied. Can either be "add" or "remove}

\item{element}{The type of the element to be modified. Can either be "node", "edge", or "layer"}

\item{name}{The name of the element to be modified. This argument is only mandatory for nodes and edges}

\item{layername}{The layer name. This argument is only mandatory for action "add" and element "node"}

\item{V1}{The start node of an edge. This argument is only mandatory for element "edge"}

\item{V2}{The end node of an edge. This argument is only mandatory for element "edge"}

\item{attributes}{The named list of attributes of the element. This argument is required only for action "add". It is optional for both elements "node" and "edge", but mandatory if the edge alread exists}

\item{multi}{A boolean whether to select multi-edges or not. This is only mandatory for action "remove" and element "edge". By default set to FALSE, in which case the attributes of the specified edge should be given}

\item{trans}{A boolean whether to add transitive edges upon removal of nodes or layers}
}
\value{

}
\description{
Track a modification of a graph
}
\examples{
g=mully:::demo()
view=pathwayView(g,"View1")
view=addStep(view,"remove","layer","")
}
