% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{charSort}
\alias{charSort}
\title{charSort}
\usage{
charSort(x)
}
\arguments{
\item{x}{A vector of strings}
}
\value{
The vector sorted alphabetically a->z
}
\description{
This function sorts a character vector
adpated from: https://stackoverflow.com/questions/38654019/rcpp-sort-descending
}
