% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR_DE.R
\name{compute_cell_type_specific_DE}
\alias{compute_cell_type_specific_DE}
\title{Cell-type specific DE}
\usage{
compute_cell_type_specific_DE(pseudobulks, design_matrix, fdr = 0.05)
}
\arguments{
\item{pseudobulks}{a pseudobulk matrix created from get_pseudobulks}

\item{design_matrix}{a design matrix created with model.matrix}

\item{fdr}{False discovery rate threshold used to filter DE genes}
}
\value{
a list of dataframes containing DE results for each cell-type
}
\description{
Calculated differential expression for each cell-type given an appropriate design matrix
}
\details{
Uses edgeR (QLR test) to calculate differential expression for each coefficient in the design matrix for each cell-type. Results are filtered based on the provided FDR threshold.
}
\examples{
example_celltype_pseudobulks <- generate_synthetic_pseudobulks()
conditions <- sapply(strsplit(colnames(example_celltype_pseudobulks), "[_-]"),function(x){x[[2]]})
names(conditions) <- colnames(example_celltype_pseudobulks)
design <- model.matrix(~conditions)
de <- compute_cell_type_specific_DE(example_celltype_pseudobulks, design)
}
