% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR_DE.R
\name{compute_cell_type_specific_DE_parallel}
\alias{compute_cell_type_specific_DE_parallel}
\title{Cell-type specific DE - parallel}
\usage{
compute_cell_type_specific_DE_parallel(
  pseudobulks,
  design_matrix,
  fdr = 0.05,
  n.cores = 1
)
}
\arguments{
\item{pseudobulks}{a pseudobulk matrix created from get_pseudobulks}

\item{design_matrix}{a design matrix created with model.matrix}

\item{fdr}{False discovery rate threshold used to filter DE genes}

\item{n.cores}{Number of CPUs to use for parallel computing}
}
\value{
a list of dataframes containing DE results for each cell-type
}
\description{
Calculated differential expression for each cell-type given an appropriate design matrix
}
\details{
Uses edgeR (QLR test) to calculate differential expression for each coefficient in the design matrix for each cell-type. Results are filtered based on the provided FDR threshold. Performs differential expression for each cell-type in parallel to speed up computations.
}
\examples{
example_celltype_pseudobulks <- generate_synthetic_pseudobulks()
conditions <- sapply(strsplit(colnames(example_celltype_pseudobulks), "[_-]"),function(x){x[[2]]})
names(conditions) <- colnames(example_celltype_pseudobulks)
design <- model.matrix(~conditions)
de <- compute_cell_type_specific_DE_parallel(example_celltype_pseudobulks, design)
}
