% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrichments.R
\name{condense_terms_multi}
\alias{condense_terms_multi}
\title{Condense Terms Multi}
\usage{
condense_terms_multi(out_list, equivalient = 0.5, verbose = FALSE)
}
\arguments{
\item{out_list}{a list of pathway enrichments for multiple DE tests obtained from do_ora or do_fgsea}

\item{verbose}{boolean, whether to print data table used as the basis for selecting representative terms.}

\item{equivalent}{a scalar value for the equivalence threshold. (see: \code{get_overlaps})}
}
\value{
The same structured list of pathway enrichments but with redundant terms removed.
}
\description{
Removes synonymous pathways from multiple sets of output pathway enrichments
}
\details{
Combines multiple sets of pathways enrichments for different cell-types / conditions 
that have not yet been condensed. Then finds overlaps between all pathway enrichments. 
Finally identifies a since consensus pathway to represent each cluster of synonymous terms.

Consensus pathway is defined as: ....
See: \code{cluster_overlaps} for details of pathway clustering.
See: \code{get_oberlaps} for deatils of overlap calculation.
}
\examples{
paths <- convert_GSEAObj_to_list(get_pathways("test"))
rich <- do_ora(paths[[1]], paths, background=unique(unlist(paths)))
rich <- condense_terms(rich)
}
