% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hierarchical_Pathways.R
\name{convert_GOBP}
\alias{convert_GOBP}
\title{Convert Gene2GO into pathways list}
\usage{
convert_GOBP(annotation_table, gene_col = "GENE", anno_col = "GO")
}
\arguments{
\item{annotation_table}{table of gene annotations from 'load_GOBP_info'}

\item{gene_col}{the name of the column of the table with geneID}

\item{anno_col}{the name of the column of the table with the GO term ID}
}
\value{
a list of GO pathways with their associated genes
}
\description{
Converts the GOBP annotations used for simulations into appropriate format for use in do_ora and do_gsea
}
