% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Pathways.R
\name{convert_GSEAObj_to_list}
\alias{convert_GSEAObj_to_list}
\title{Convert GSEAObj}
\usage{
convert_GSEAObj_to_list(gsea_obj)
}
\arguments{
\item{gsea_obj}{a GeneSet or GeneSetCollection object, such as those obtained from the msigdb database.}
}
\value{
a list of gene sets.
}
\description{
Converts the GSEA object obtained from msigdb or other Bioconductor pathway database to a list suitable for use in `do_ora`.
}
\examples{
genesets <- get_pathways("test")
geneset_list <- convert_GSEAObj_to_list(genesets)
length(geneset_list)
}
