% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hierarchical_Pathways.R
\name{create_rand_pathway_sim}
\alias{create_rand_pathway_sim}
\title{Generate Random Pathways Simulation}
\usage{
create_rand_pathway_sim(
  n_replicas,
  n_path,
  n_genes_per_path,
  fpr,
  seed = 2938,
  max_term_size = 1000,
  gene2terms
)
}
\arguments{
\item{n_replicas}{number of simulations to generate}

\item{n_path}{number of pathways selected per simulation}

\item{n_genes_per_path}{number of genes significantly DE for each selected pathway}

\item{fpr}{proportion of DE genes that are false-positives}

\item{seed}{the random seed to ensure reproducibility}

\item{max_term_size}{the maximum genes per pathway to consider that pathway}

\item{gene2terms}{table of gene->GOterm annotations (from: load_GOBP_info())}
}
\value{
a list of simulation output, each simulation output is a list that contains:
gt_terms : the truly DE GO terms (vector)
de_genes : the DE genes including false-positives (vector), needed for do_ora
background : the tissue-specific genes used as the background (vector), needed for do_ora
gsea_input : named vector of fold changes with the de_genes at the top, needed for do_gsea
}
\description{
Generates DE gene output for ORA or GSEA for randomly selected pathways.
}
\details{
Combines multiple functions to first randomly select a tissue for each simulation ('get_tissue_specific_genes'), select GO:BP terms ('select_random_terms'), randomly draw genes from those terms ('draw_genes'), generate log2fcs which are identical for all replicates ('generate_log2fcs'), and create the gsea input ('generate_gsea_input'). 

This function is for the random pathways case, where terms are drawn at random equally likely across all valid pathways. 

Constraints imposed on pathways are: tissue-specific gene expression, sufficient genes for the number of genes per pathway required, and no more than 'max_term_size' genes. Gene number constraints are imposed after filtering for tissue-specific gene expression.
}
\examples{
gene2go <- load_GOBP_info()[1:50000,] # reduced to limit runtime
sims <- create_rand_pathway_sim(1, 5, 15, fpr=0.05, gene2terms=gene2go)
paths <- convert_GOBP(gene2go)
ora_out <- do_ora(sims[[1]]$de_genes,pathways=paths, background=sims[[1]]$background)
gsea_out <- do_gsea(sims[[1]]$gsea_input, pathways=paths, nperm=100) # reduced number of permutations to limit runtime 
}
