% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reactome.R
\name{downloadPathway}
\alias{downloadPathway}
\title{Download Reactome Pathways in BioPAX level 2 and 3}
\usage{
downloadPathway(pathwayID, biopaxLevel = "3", destDirectory, overwrite = F)
}
\arguments{
\item{pathwayID}{The Reactome ID or list of IDs of the pathways to be downloaded. The ID should start with R-HSA-.}

\item{biopaxLevel}{The BioPAX Level, 2 or 3. By default set to 3.}

\item{destDirectory}{The Directory in which the Pathway Files should be saved. If missing, the files are saved in the working directory. The Reactome IDs are used to name the files.}

\item{overwrite}{A Boolean whether to overwrite existing files with the same name.}
}
\value{
The Directory in which the files are saved.
}
\description{
Download Reactome Pathways in BioPAX level 2 and 3
}
\examples{
\dontrun{ 
downloadPathway(c("R-HSA-195721","R-HSA-9609507"),biopaxLevel=3,overwrite=T)
}
}
