% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG.R
\name{filterGeneList}
\alias{filterGeneList}
\title{Remove invalid Gene IDs from a gene list}
\usage{
filterGeneList(geneList)
}
\arguments{
\item{geneList}{A list of given KEGG Gene IDs}
}
\value{
geneList A list of filtered valid gene IDs
}
\description{
Remove invalid Gene IDs from a gene list
}
\examples{
geneList=c("hsa:10458", "hsa:108", "hsa:111", "hsa:112", "hsa:113", "hsa:114")
filteredGeneList=filterGeneList(geneList)
}
\author{
Zaynab Hammoud
}
