% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hierarchical_Pathways.R
\name{generate_gsea_input}
\alias{generate_gsea_input}
\title{Generate GSEA input}
\usage{
generate_gsea_input(true_genes, genes, log2fcs = NULL, seed = NULL)
}
\arguments{
\item{true_genes}{true-positive genes to be placed at the top.}

\item{genes}{background genes.}

\item{log2fcs}{vector of log2fcs, if NULL log2fcs are generated using 'generate_log2fcs'}

\item{seed}{the random seed to set.}
}
\value{
a named vector of log2fcs with the true genes at the top (most positive log2fcs).
}
\description{
Generates a vector of log2fcs for GSEA with a set of true genes at the top.
}
\details{
Creates a vector of weighted genes using provided log2fcs or generating these values using 'generate_log2fcs' if not provided. Then assigns the top highest positive values to the provided 'true_genes' after randomly permuting them, the rest of the values are assigned to the background genes provided in 'genes' which are also permuted to ensure no possible patterns. If there are insufficient log2fcs provided, the missing values are filled in with 0s.
}
\examples{
tissue_specific <- get_tissue_specific_genes(1, seed=301)
TP_genes <- sample(tissue_specific[[1]], 50)
gsea_input <- generate_gsea_input(TP_genes, tissue_specific[[1]], seed=2911) 
gsea_input <- generate_gsea_input(TP_genes, tissue_specific[[1]], log2fcs=rnorm(2000), seed=2911) 
}
