% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{generate_heatmap_data}
\alias{generate_heatmap_data}
\title{Collect Heatmap Data}
\usage{
generate_heatmap_data(list_of_rich, pathways)
}
\arguments{
\item{list_of_rich}{a list of output from do_ora or do_fgsea typically from multiple cell-types}

\item{pathways}{which pathways to aggregate.}
}
\value{
a list of "scores" and "pvalues" for the provided pathways across all cell-types.
}
\description{
Collects enrichment scores and FDR p-values from a set of enrichments.
}
\details{
Combines enrichment scores and FDRs from a list of enrichment output into two matrices which can be plotted as a heatmap across all cell-types.
pathways missing from one or another cell-type are assigned scores of 0, and p-values of 1.
}
\examples{
list_of_rich <- lapply(rpois(5, lambda=10), generate_synthetic_enrichments, ngenes=100)
names(list_of_rich) <- paste("celltype", 1:length(list_of_rich), sep="")
heat_data <- generate_heatmap_data(list_of_rich, pathways=c("pathway1", "pathway2", "pathway5", "pathway7"))
dim(heat_data$scores)
dim(heat_data$pvalues)
}
