% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hierarchical_Pathways.R
\name{generate_log2fcs}
\alias{generate_log2fcs}
\title{Generate generate log2fcs}
\usage{
generate_log2fcs(seed = NULL)
}
\arguments{
\item{seed}{the random seed.}
}
\value{
a vector of log2 fold changes.
}
\description{
Generates example log2fcs from GTEX data.
}
\details{
Randomly selects two tissues from the GTEX database, and calculates the log2fc for every gene betweeen them. Excludes any NAs or infinite values resulting from a gene being 0 in the denominator.
}
\examples{
l2fc <- generate_log2fcs(seed=1938)
}
