% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_Examples.R
\name{generate_synthetic_enrichments}
\alias{generate_synthetic_enrichments}
\title{Generate synthetic enrichments}
\usage{
generate_synthetic_enrichments(npathways = 10, ngenes = 200)
}
\arguments{
\item{npathways}{number of pathway enrichments to generate}

\item{ngenes}{number of total genes to use to populate enrichments.}
}
\value{
A list containing a dataframe of the same format as the output of do_ora or do_fgsea and a list of contributing genes.
}
\description{
Generates a synthetic enrichments for testing other functions
}
\details{
Generates a set of random output formatted to look like the output from do_ora or do_fgsea.
}
\examples{
example_enrichments <- generate_synthetic_enrichments(10)
dim(example_enrichments$results)
length(example_enrichments$contrib)
}
