% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_Examples.R
\name{generate_synthetic_pseudobulks_one_cell_type}
\alias{generate_synthetic_pseudobulks_one_cell_type}
\title{Generate synthetic pseudobulks (one-celltype)}
\usage{
generate_synthetic_pseudobulks_one_cell_type(
  ngenes = 100,
  de_prop = 0.5,
  nconditions = 2,
  nsamples_per_condition = 3,
  dispersion_factor = 0.2
)
}
\arguments{
\item{ngenes}{number of genes to generate data for.}

\item{de_prop}{proportion of genes to be differentially expressed between biological conditions.}

\item{nconditions}{number of different biological conditions for which to generate synthetic data.}

\item{nsamples_per_condition}{number of samples for each biological condition.}

\item{dispersion_factor}{dispersion parameter (1/size) of the negative bionomial distribution.}
}
\value{
A matrix of pseudobulk expression with columns named as: [celltype]_[condition]-[sample]
}
\description{
Generates a synthetic pseudobulk matrix for testing purposes.
}
\details{
Generates synthetic pseudobulk data for one cell-type. Sample-specific library sizes are drawn from an exponential distribution and rescaled to the median value. Gene-specific means are drawn from an exponential distribution with rate=1/100, counts are drawn from a negative binomial distribution.
}
\examples{
example_celltype_pseudobulks <- generate_synthetic_pseudobulks_one_cell_type()
dim(example_celltype_pseudobulks)
}
