% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_Examples.R
\name{generate_test_cellcounts}
\alias{generate_test_cellcounts}
\title{Generate Cell Counts for testing}
\usage{
generate_test_cellcounts()
}
\value{
a list with items: counts = umi count matrix, donors = vector of sample of origin labels, celltype = vector of celltype labels
}
\description{
Generates a synthetic gene x cell umi count matrix for testing methods.
}
\details{
Generates a matrix of uniform Poisson distributed counts, as well as vectors of cell-type
and donor of origin labels. These data include one cell-type represented by 1 cell per donor, 
one with 15 cells/donor and one with 30 cells/donor to check for errors in methods.
}
\examples{
example_data <- generate_test_cellcounts()
dim(example_data$counts)
table(example_data$donors)
table(example_data$celltypes)
}
