% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Pathways.R
\name{get_pathways}
\alias{get_pathways}
\title{Get Pathways from MSigDb}
\usage{
get_pathways(species = c("mouse", "human"), include.celltype = FALSE)
}
\arguments{
\item{species}{whether to download human or mouse pathways}

\item{include.celltype}{whether to include cell-type marker genesets.}
}
\value{
GeneSetCollection object of pathways.
}
\description{
Downloads common pathways from msigdb for either human or mouse.
}
\details{
Downloads MSigDB Hallmark pathways, GO biological processes, Reactome and Biocarta pathways from `msigdb` for either human or mouse. For human it also adds the KEGG pathways. Optionally also downloads cell-type marker genesets.
}
\note{
requires dbplyr version 3.2.4 or earlier
}
\examples{
#genesets <- get_pathways("mouse")
#genesets <- get_pathways("human")
}
