% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hierarchical_Pathways.R
\name{get_tissue_specific_genes}
\alias{get_tissue_specific_genes}
\title{Obtain tissue-specific genes}
\usage{
get_tissue_specific_genes(n, seed = NULL, threshold = 5)
}
\arguments{
\item{n}{the number of tissues to randomly sample.}

\item{seed}{set a random seed for the sampling.}

\item{threshold}{the cpm threshold to define a gene as expressed in the tissue.}
}
\value{
a list of tissue specific genes for each selected tissue.
}
\description{
Get the tissue specific genes for a random selection of tissues.
}
\details{
Randomly selects "n" tissues and gets the a list of genes with an average expression > threshold counts per million (cpm).
}
\examples{
genes <- get_tissue_specific_genes(3, seed=3819, threshold=5)
}
