% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniProtKB.R
\name{getAllUPKB}
\alias{getAllUPKB}
\title{Get all proteins' entries from UniProt}
\usage{
getAllUPKB(up)
}
\arguments{
\item{up}{The UniProt.ws Object}
}
\value{
a dataframe containing the Protein's entries with the ID and Name
}
\description{
Get all proteins' entries from UniProt
}
\note{
Should be preceded by UniProt.ws() to get the UniProt.ws Object
}
\examples{
up=UniProt.ws()
allProteins=getAllUPKB(up)
}
