% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugBank.R
\name{getDBCarriers}
\alias{getDBCarriers}
\title{Get the carrier proteins involved in movement of given drugs across biological membranes}
\usage{
getDBCarriers(data, drugList)
}
\arguments{
\item{data}{The dataframe containing the parsed information of DrugBank. This argument can be obtained using the function loadDBXML(DrugBankFile)}

\item{drugList}{The list of DrugBank Ids of the drugs. This argument can be either a string (one drug) or a list of strings (multiple drugs)}
}
\value{
A dataframe containing all information on the carrier proteins involved in movement of the given drugs across biological membranes
}
\description{
Get the carrier proteins involved in movement of given drugs across biological membranes
}
