% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugBank.R
\name{getDBDrug}
\alias{getDBDrug}
\title{Get DrugBank drug entry}
\usage{
getDBDrug(data, drug)
}
\arguments{
\item{data}{The dataframe containing the parsed information of DrugBank. This argument can be obtained using the function loadDBXML(DrugBankFile)}

\item{drug}{The ID or list of IDs of the DrugBank drug entries starting with "DB"}
}
\value{
A dataframe containing the DrugBank entry with its information
}
\description{
Get DrugBank drug entry
}
\examples{
\dontrun{ 
data=loadDBXML(DrugBankFilePath)
getDBDrug(data, "DB00001")
}
}
