% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugBank.R
\name{getDBDrugInteractions}
\alias{getDBDrugInteractions}
\title{Get DrugBank Drug to Drug Interactions}
\usage{
getDBDrugInteractions(data, drug)
}
\arguments{
\item{data}{The dataframe containing the parsed information of DrugBank. This argument can be obtained using the function loadDBXML(DrugBankFile)}

\item{drug}{The ID of the DrugBank drug entry starting with "DB". This argument can be either a string (one drug) or a list of strings (multiple drugs).}
}
\value{
A dataframe containing the DrugBank interactions in which the given drug is involved
}
\description{
Get DrugBank Drug to Drug Interactions
}
\examples{
getDBDrugInteractions(data,"DB06605")
}
