% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugBank.R
\name{getDBEnzymes}
\alias{getDBEnzymes}
\title{Get the enzymes inhibited/induced or involved in metabolism by given DrugBank drugs}
\usage{
getDBEnzymes(data, drugList)
}
\arguments{
\item{data}{The dataframe containing the parsed information of DrugBank. This argument can be obtained using the function loadDBXML(DrugBankFile)}

\item{drugList}{The list of DrugBank Ids of the drugs. This argument can be either a string (one drug) or a list of strings (multiple drugs)}
}
\value{
A dataframe containing all information on the enzymes inhibited/induced or involved in metabolism by the given drug list
}
\description{
Get the enzymes inhibited/induced or involved in metabolism by given DrugBank drugs
}
