% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugBank.R
\name{getDBTargets}
\alias{getDBTargets}
\title{Get the targets of given DrugBank drugs}
\usage{
getDBTargets(data, drugList)
}
\arguments{
\item{data}{The dataframe containing the parsed information of DrugBank. This argument can be obtained using the function loadDBXML(DrugBankFile)}

\item{drugList}{The list of DrugBank Ids of the drugs. This argument can be either a string (one drug) or a list of strings (multiple drugs)}
}
\value{
A dataframe containing all information on the tragets of the given drug list
}
\description{
Get the targets of given DrugBank drugs
}
