% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intraedges.R
\name{getDBtoUPKB}
\alias{getDBtoUPKB}
\title{Get DrugBank Drugs to UniProt Proteins Relations from DrugBank}
\usage{
getDBtoUPKB(data, drugList, proteinList)
}
\arguments{
\item{data}{The dataframe containing the parsed information of DrugBank. This argument can be obtained using the function loadDBXML(DrugBankFile)}

\item{drugList}{The list of DrugBank Ids of the drugs. This argument can be either a string (one drug) or a list of strings (multiple drugs)}

\item{proteinList}{The list of UniProt Ids of the proteins}
}
\value{
A dataframe containing the connections between DrugBank drugs and UniProt proteins retrieved from DrugBank
}
\description{
Get DrugBank Drugs to UniProt Proteins Relations from DrugBank
}
\examples{
getDBtoUPKB(data,c("DB00001","DB00002","DB00006"),c("P02747","P00734","P07204","P05164"))
}
