% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reactome.R
\name{getInternalIDs}
\alias{getInternalIDs}
\title{Get Internal ID of a given node or list of nodes}
\usage{
getInternalIDs(g, nodes = NULL)
}
\arguments{
\item{g}{The graph generated from the given biopax object.}

\item{nodes}{The list of external IDs of the nodes. If missing, the list of all internal IDs will be returned.}

\item{biopax}{The biopax object}
}
\value{
A dataframe with the mappings between the external and internal  IDs
}
\description{
Get Internal ID of a given node or list of nodes
}
\examples{
\dontrun{ 
biopax=readBiopax(pi3k.owl)
g=pathway2Mully(biopax,"pathway1")
getInternalIDs(wntBiopax,c("R-HSA-5368514","Q9BQB4"))
}
}
