% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG.R
\name{getKeggGene}
\alias{getKeggGene}
\title{Get Kegg gene}
\usage{
getKeggGene(geneList)
}
\arguments{
\item{geneList}{List of Genes as formatted in KEGG Genes}
}
\value{
A data frame of kegg data
}
\description{
Get Kegg gene
}
\examples{
geneList=c("hsa:122706","hsa:4221","hsa:8312")
genes=  getKeggGene(geneList)

}
