% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG.R
\name{getKeggGeneParallel}
\alias{getKeggGeneParallel}
\title{Parallel processing of KEGG entries}
\usage{
getKeggGeneParallel(geneList, chunkSize = 10)
}
\arguments{
\item{geneList}{List of Genes as formatted in KEGG Genes}

\item{chunkSize}{Size of each chunk (default: 10, as per KEGGREST limit)}
}
\value{
A data frame of KEGG data
}
\description{
Parallel processing of KEGG entries
}
\examples{
geneList = c("hsa:122706", "hsa:4221", "hsa:8312")
genes = getKeggGeneParallel(geneList)
}
\author{
Mohammad Al Maaz
}
