% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intraedges.R
\name{getKeggOmimRelation}
\alias{getKeggOmimRelation}
\title{Gets the genes and diseases that are referenced to each other. Genes are extracted from the graph}
\usage{
getKeggOmimRelation(g, biopax)
}
\arguments{
\item{g}{A Mully Graph Object}

\item{biopax}{A Biopax Object}
}
\value{
A Dataframe with kegggenes ids and omim ids
}
\description{
Gets the genes and diseases that are referenced to each other. Genes are extracted from the graph
}
\note{
must be preceded by addGenesLayer(g,biopax) function
}
\examples{
biopax=readBiopax("wnt.owl")
pathwayID=listPathways(biopax)$id[1]
g=Multipath::pathway2Mully(biopax,pathwayID)
g=addGenesLayer(g,biopax)
dataframe = getKeggOmimRelation(g,biopax)
}
\author{
Mohammad Al Maaz
}
