% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intraedges.R
\name{getKEGGtoDATABASE}
\alias{getKEGGtoDATABASE}
\title{Get KEGG Genes to a Database cross-reference  from KEGG}
\usage{
getKEGGtoDATABASE(dbName, geneList)
}
\arguments{
\item{dbName}{Name of database to find as string- only 1 input allowed}

\item{geneList}{List of Proteins as formatted in UPKB}
}
\value{
Dataframe of the Genes-Proteins relation
}
\description{
Get KEGG Genes to a Database cross-reference  from KEGG
}
\examples{
geneList=c("hsa:122706","hsa:4221","hsa:8312")
keggToDb = getKEGGtoDATABASE("UniProt",geneList)
}
\author{
Mohammad Al Maaz
}
