% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intraedges.R
\name{getKEGGtoOMIM}
\alias{getKEGGtoOMIM}
\title{Get KEGG Genes to Omim relations from KEGG}
\usage{
getKEGGtoOMIM(geneList)
}
\arguments{
\item{geneList}{List of Genes as formatted in KEGG Genes}
}
\value{
Dataframe of the Genes-Disease relation
}
\description{
Get KEGG Genes to Omim relations from KEGG
}
\examples{
geneList=c("hsa:122706","hsa:4221","hsa:8312")
kEGG2OMIM = getKEGGtoOMIM(geneList)
}
\author{
Mohammad Al Maaz
}
