% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG.R
\name{getKEGGtoUPKB}
\alias{getKEGGtoUPKB}
\title{Get KEGG Genes to UniProt relations from KEGG}
\usage{
getKEGGtoUPKB(geneList, proteinList)
}
\arguments{
\item{geneList}{List of Genes as formatted in KEGG Genes}

\item{proteinList}{List of Proteins as formatted in UPKB}
}
\value{
Dataframe of the Genes-Proteins relation
}
\description{
Get KEGG Genes to UniProt relations from KEGG
}
\examples{
proteinList = c("P02747","P00734","P07204","A0A0S2Z4R0","O15169")
geneList=c("hsa:122706","hsa:4221","hsa:8312")
kEGG2UPKB = getKEGGtoUPKB(geneList,proteinList)
}
