% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intraedges.R
\name{getKeggUpkbRelations}
\alias{getKeggUpkbRelations}
\title{Gets the genes and proteins that are referenced to each other}
\usage{
getKeggUpkbRelations(g, biopax)
}
\arguments{
\item{g}{A Mully graph object}

\item{biopax}{A biopax object}
}
\value{
Dataframe of related proteins and genes
}
\description{
Gets the genes and proteins that are referenced to each other
}
\examples{
up = UniProt.ws()
biopax=readBiopax("wnt.owl") 
pathwayID=listPathways(biopax)$id[1]
g=Multipath::pathway2Mully(biopax,pathwayID) #This is a mully graph that contains a protein layer
g=addGenesLayer(g,biopax)
UPKB2KEGG = getUPKBtoKEGG(g, biopax)
}
\author{
Mohammad Al Maaz
}
