% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intraedges.R
\name{getOmimToKEGG}
\alias{getOmimToKEGG}
\title{Get Omim to KEGG Genes relations from OMIM}
\usage{
getOmimToKEGG(omimIds)
}
\arguments{
\item{omimIds}{list of OMIM Ids}
}
\value{
Dataframe of the Genes-OMIM relation
}
\description{
Get Omim to KEGG Genes relations from OMIM
}
\note{
should be preceded by calling  romim::set_key('KEY'). The KEY could be requested via omim's official website.
}
\examples{
library(romim)
romim::set_key('KEY')
kEGG2OMIM = getOmimToKEGG(c("611137", "613733", "603816"))
}
\author{
Mohammad Al Maaz
}
