% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reactome.R
\name{getPathwayID}
\alias{getPathwayID}
\title{Get internal pathway ID in a BioPAX file}
\usage{
getPathwayID(biopax, reactomeID)
}
\arguments{
\item{biopax}{The biopax object}

\item{reactomeID}{The Reactome ID of the pathway}
}
\value{
The internal ID of the pathway in the parsed BioPAX object
}
\description{
Get internal pathway ID in a BioPAX file
}
\note{
This should be preceded by readBiopax(filepath) to obtain the biopax object
}
\examples{
biopax=readBiopax("pi3k.owl")
id=getPathwayID(biopax,"R-HSA-167057")
pi3kmully=pathway2mully(biopax,id)
}
