% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG.R
\name{getRelatedGenes}
\alias{getRelatedGenes}
\title{Get proteins that has a reference to KEGG Genes from a mully graph and a biopax}
\usage{
getRelatedGenes(g, biopax)
}
\arguments{
\item{g}{A Mully graph}

\item{biopax}{A biopax file}
}
\value{
List of Protein
}
\description{
Get proteins that has a reference to KEGG Genes from a mully graph and a biopax
}
\examples{
downloadPathway("R-HSA-195721",biopax=3) 
biopax=readBiopax("R-HSA-195721.owl") 
pathwayID=listPathways(biopax)$id[1] 
g=pathway2Mully (biopax, pathwayID) 
relatedGenes = getRelatedGenes(g,biopax)
}
