% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intraedges.R
\name{getUPKBDBRelations}
\alias{getUPKBDBRelations}
\title{Get Protein and Drugs relations from UniProt and DrugBank}
\usage{
getUPKBDBRelations(data, proteinList, drugList)
}
\arguments{
\item{data}{The dataframe containing the parsed information of DrugBank. This argument can be obtained using the function loadDBXML(DrugBankFile)}

\item{proteinList}{The list of UniProt Ids of the proteins}

\item{drugList}{The list of DrugBank Ids of the drugs. This argument can be either a string (one drug) or a list of strings (multiple drugs)}
}
\value{
A dataframe containing the connections between DrugBank drugs and UniProt proteins retrieved from DrugBank and UniProt
}
\description{
Get Protein and Drugs relations from UniProt and DrugBank
}
\note{
Should be preceded by:
loadDBXML(DrugBankFile) to get the argument data
}
\examples{
\dontrun{
data=readDBXML(DBXMLFilePath)
relations=getUPKBDBRelations(data,c("P02747","P05164"),c("DB00001","DB00006"))
}
}
