% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniProtKB.R
\name{getUPKBInfo}
\alias{getUPKBInfo}
\title{Get Proteins from UniProtKB}
\usage{
getUPKBInfo(up, proteins, col)
}
\arguments{
\item{up}{The UniProt.ws Object}

\item{proteins}{The list of UniProtKB Proteins ID to be retrieved}

\item{col}{The list of attributes associated to the UniProtKB Entries to be retrieved}
}
\value{
a dataframe containing the protein entries with the selected attributes
}
\description{
Get Proteins from UniProtKB
}
\note{
Should be preceded by UniProt.ws() to get the UniProt.ws Object
To get the list of possible columns, you can call columns(UniProt.ws())
}
\examples{
\dontrun{ 
up <- UniProt.ws()
getUPKBInfo(up,c("Q6ZS62","P14384"),c("protein_name","go")) 
} 
}
