% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMIM.R
\name{getUPKBRelatedDiseases}
\alias{getUPKBRelatedDiseases}
\title{Get proteins that has a reference to OMIM from a mully graph and a biopax}
\usage{
getUPKBRelatedDiseases(g, biopax)
}
\arguments{
\item{g}{A Mully graph}

\item{biopax}{A biopax file}
}
\value{
Dataframe of Proteins-OMIM relation
}
\description{
Get proteins that has a reference to OMIM from a mully graph and a biopax
}
\note{
should be preceded by calling  romim::set_key('KEY'). The KEY could be requested via omim's official website.
}
\examples{
\dontrun{ 
biopax=readBiopax("wnt.owl") 
pathwayID=listPathways(biopax)$id[1]
g=Multipath::pathway2Mully(biopax,pathwayID) #This is a mully graph that contains a protein layer
getUPKBRelatedDiseases(g,biopax)
}
}
\author{
Mohammad Al Maaz
}
