% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intraedges.R
\name{getUPKBtoKEGG}
\alias{getUPKBtoKEGG}
\title{Get UniProt Proteins to KEGG Genes relations from UPKB}
\usage{
getUPKBtoKEGG(proteinList, geneList)
}
\arguments{
\item{proteinList}{List of Proteins as formatted in UPKB}

\item{geneList}{List of Genes as formatted in KEGG Genes}
}
\value{
Dataframe of the Proteins-Genes relation
}
\description{
Get UniProt Proteins to KEGG Genes relations from UPKB
}
\examples{
proteinList = c("P02747","P00734","P07204","A0A0S2Z4R0","O15169")
geneList=c("hsa:122706","hsa:4221","hsa:8312")
UPKB2KEGG = getUPKBtoKEGG(geneList,proteinList)
}
\author{
Mohammad Al Maaz, Zaynab Hammoud
}
