% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hierarchical_Pathways.R
\docType{data}
\name{GOBP_layers}
\alias{GOBP_layers}
\title{Preprocessed Gene Ontology: GOBP_layers}
\format{
a list of 18 vectors:
\describe{
   \item{item name}{layer within the GO:BP ontology starting with 1 = root.}
   \item{value}{a vector of GO term ids for the terms that are reachable in exactly that many steps from the root.}
 }
}
\source{
{GO.db} Bioconductor database package.
}
\usage{
GOBP_layers
}
\description{
Reorganization of the GO:BP ontology into a stack of layers at different depths from the root term. Individual GO terms appear in multiple layers if there are multiple routes of different length from the root to that term. Created: May 2025
}
\keyword{datasets}
