% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pseudobulks.R
\name{group_rowmeans}
\alias{group_rowmeans}
\title{Group Row Means}
\usage{
group_rowmeans(MAT, group_labs, type = c("mean", "sum"))
}
\arguments{
\item{MAT}{a matrix or sparse matrix of values to be aggregated}

\item{group_labs}{a vector of group labels equal in length to the number of columsn of MAT.}

\item{type}{whether to calculate the row means or row sums}
}
\value{
a matrix of the row means for each group.
}
\description{
Optimized code for calculating the row means or row sums of a matrix after grouping the columns by a vector of groups.
}
\details{
Uses my_rowSums and my_rowMeans to calculate a matrix of row means or row sums for each group devinfed in group_labs. Note normalized expression should be averaged while raw counts should be summed.
}
\examples{
example_data <- generate_test_cellcounts()
avg_celltype_counts <- group_rowmeans(example_data$counts,example_data$celltypes)
avg_donor_counts <- group_rowmeans(example_data$counts,example_data$donors)
}
