% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hierarchical_Pathways.R
\docType{data}
\name{gtex_genes}
\alias{gtex_genes}
\title{GTEx Tissue-specific Expression: gtex_genes}
\format{
a matrix containing 20801 rows and 30 columns:
\describe{
   \item{columns}{correspond to 30 different tissues in alphabetical order: Adipose, Adrenal, Bladder, Blood, Blood Vessel, Brain, Breast, Cervix, Colon, Esophagus, Fallopian Tube, Heart, Kidney, Liver, Lung, Muscle, Nerve, Ovary, Pancreas, Pituitary, Prostate, Salivary, Skin, Small Intestine, Spleen, Stomach, Testis, Thyroid, Uterus, Vagina.}
   \item{rows}{genes, contains a mix of gene symbols and Ensembl IDs(if no gene symbol was available)}
 }
}
\source{
\url{https://storage.googleapis.com/adult-gtex/bulk-gex/v10/rna-seq/GTEx_Analysis_v10_RNASeQCv2.4.2_gene_reads.gct.gz}
}
\usage{
gtex_genes
}
\description{
Average expression (counts-per-million) as measured by bulkRNAseq in each tissue from the GTEx database (v10, RNASeqQCv2.4.2). Created: Jun 2025
}
\keyword{datasets}
