% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG.R
\name{interactionKegg}
\alias{interactionKegg}
\title{Interaction between KEGG genes and other databases}
\usage{
interactionKegg(keggInput)
}
\arguments{
\item{keggInput}{data frame of all data retrieved from kegg genes database}
}
\value{
A data frame of each gene vs interactions with other databases
}
\description{
Interaction between KEGG genes and other databases
}
\examples{
geneList=c("hsa:122706","hsa:4221","hsa:8312")
genes = getKeggGene(geneList)
genesinteraction = interactionKegg(genes)
}
\author{
Mohammad Al Maaz
}
