% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{intersectToList}
\alias{intersectToList}
\title{intersectToList}
\usage{
intersectToList(lt, x)
}
\arguments{
\item{lt}{A list of vectors of strings}

\item{x}{A vector of strings}
}
\description{
This function intersects a list of string vectors with a single string vector.
adpated from: https://jokergoo.github.io/2023/04/05/speed-up-over-representation-enrichment-analysis/
originally created by: Zuguang Gu
}
