% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG.R
\name{is_valid_kegg_gene}
\alias{is_valid_kegg_gene}
\title{Check KEGG Gene Code Validity}
\usage{
is_valid_kegg_gene(gene_code)
}
\arguments{
\item{gene_code}{A KEGG Gene ID}
}
\value{
boolean True if exists, false if not
}
\description{
Check KEGG Gene Code Validity
}
\examples{
is_valid_kegg_gene("hsa:108")
}
\author{
Zaynab Hammoud
}
