% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hierarchical_Pathways.R
\name{load_GOBP_info}
\alias{load_GOBP_info}
\title{Load the GOBP ontology annotation}
\usage{
load_GOBP_info()
}
\value{
a table with the following columnd: the ENTREZ gene id, the go term annotation, the evidence code, the Ontology (BP), and the gene symbol (GENE).
}
\description{
This load the GO Biological Process annotations for all human genes using the org.hs.eg.db package.
}
\examples{
gene2gobp <- load_GOBP_info()
}
