% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugBank.R
\name{loadDBXML}
\alias{loadDBXML}
\title{Load DrugBank XML file}
\usage{
loadDBXML(file)
}
\arguments{
\item{file}{The path to the DrugBank XML file. This can be downloaded from the DrugBank official Website (drugbank.ca). An account with an institutional e-mail is required.}
}
\value{
A dataframe containing the parsed information from DrugBank. This can be used to extract any additional information on the DrugBank entries
}
\description{
Load DrugBank XML file
}
\note{
This function should be called before using any function to query the DrugBank database. Since the parsing of DrugBank takes time, this function should only be called once.
}
