% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiPath.R
\docType{package}
\name{MultiPath}
\alias{MultiPath}
\title{MultiPath}
\description{
Performs and visualizes many pathway enrichment tests, as well as pseudo-bulk differential expression for single-cell RNAseq
}
\author{
Tallulah Andrews <@tandrew6@uwo.ca>
}
