% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pseudobulks.R
\name{my_colMeans}
\alias{my_colMeans}
\title{colMeans wrapper}
\usage{
my_colMeans(x)
}
\arguments{
\item{x}{a matrix or Matrix or a vector.}
}
\value{
the column means of the matrix or the original vector if supplied with a vector
}
\description{
A wrapper for Matrix::colMeans that automatically checks if the input is a single row if so it returns just that row.
}
\examples{
example_data <- generate_test_cellcounts()
my_colMeans(example_data$counts)
my_colMeans(c(1,2,3,4,5,6,7,8,10))
}
