% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pseudobulks.R
\name{my_colVars}
\alias{my_colVars}
\title{Sparse-Matrix compatible colVars}
\usage{
my_colVars(x)
}
\arguments{
\item{x}{a matrix or Matrix or a vector.}
}
\value{
variances of each column of the matrix or 0s if supplied with a vector.
}
\description{
Uses colMeans to calculate variance of each column for sparse matrices.
}
\examples{
example_data <- generate_test_cellcounts()
my_colVars(example_data$counts)
my_colVars(c(1,2,3,4,5,6,7,8,10))
}
