% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pseudobulks.R
\name{my_rowSums}
\alias{my_rowSums}
\title{rowSums wrapper}
\usage{
my_rowSums(x)
}
\arguments{
\item{x}{a matrix or Matrix or a vector.}
}
\value{
the row sums of the matrix or the original vector if supplied with a vector
}
\description{
A wrapper for Matrix::rowSums that automatically checks if the input is a single column if so it returns just that column.
}
\examples{
example_data <- generate_test_cellcounts()
my_rowSums(example_data$counts)
my_rowSums(c(1,2,3,4,5,6,7,8,10))
}
