% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pseudobulks.R
\name{my_rowVars}
\alias{my_rowVars}
\title{Sparse-Matrix compatible rowVars}
\usage{
my_rowVars(x)
}
\arguments{
\item{x}{a matrix or Matrix or a vector.}
}
\value{
variances of each row of the matrix or 0s if supplied with a vector.
}
\description{
Uses rowMeans to calculate variance of each row for sparse matrices.
}
\examples{
example_data <- generate_test_cellcounts()
my_rowVars(example_data$counts)
my_rowVars(c(1,2,3,4,5,6,7,8,10))
}
