% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR_DE.R
\name{one_cell_type_DE_edgeR}
\alias{one_cell_type_DE_edgeR}
\title{Cell-type specific DE - one cell-type (edgeR)}
\usage{
one_cell_type_DE_edgeR(dat, design, fdr = 0.05)
}
\arguments{
\item{dat}{a pseudobulk matrix subset to a single cell-type}

\item{design}{a design matrix created with model.matrix for columns of dat}

\item{fdr}{False discovery rate threshold used to filter DE genes}
}
\value{
a dataframe of DE results for each coefficient in the design matrix.
}
\description{
Calculated differential expression for a cell-type given an appropriate design matrix using edgeR
}
\details{
Called by compute_cell_type_specific_DE to calculate DE of each cell-type. This function filters genes for those with total counts > 5 and detection in at least 3 samples. Then runs DE using edgeR on the provided data & design matrix. Normalization is performed using TMM, and DE is tested using quasi-likelihood (QLF test). Each coefficient is tested and results are combined into a single dataframe. See: edgeR::glmQLFit and glmQLRTest for details.
Due to frequent errors in TMM normalization with pseudobulk data, total counts per sample is added to the design matrix automatically.For simplicity, the DE results for total counts are not returned.
}
\examples{
example_celltype_pseudobulks <- generate_synthetic_pseudobulks()
conditions <- sapply(strsplit(colnames(example_celltype_pseudobulks), "[_-]"),function(x){x[[2]]})
cell_type <- sapply(strsplit(colnames(example_celltype_pseudobulks), "[_-]"),function(x){x[[1]]})
names(conditions) <- colnames(example_celltype_pseudobulks)
conditions <- conditions[cell_type == cell_type[1]]
dat <- example_celltype_pseudobulks[,cell_type == cell_type[1]]
design <- model.matrix(~conditions)
rownames(design) <- names(conditions)
de <- one_cell_type_DE_edgeR(dat, design)
}
