% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{overlapsListvsList}
\alias{overlapsListvsList}
\title{overlapsListvsList}
\usage{
overlapsListvsList(lt1, lt2)
}
\arguments{
\item{lt1}{A list of vectors of strings}

\item{lt2}{A list of vectors of strings}
}
\value{
out a matrix[i,j] of integers overlaps between lt1[[i]] and lt2[[j]]
}
\description{
This function intersects a list of string vectors with another list of string vectors.
Outputs the number of items in the overlap between a pair of string vectors.
Rows are vectors from lt1, columns are vectors from lt2
adpated from: https://jokergoo.github.io/2023/04/05/speed-up-over-representation-enrichment-analysis/
}
