% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reactome.R
\name{pathway2Mully}
\alias{pathway2Mully}
\title{Build a mully graph from a given pathway}
\usage{
pathway2Mully(biopax, pathwayID)
}
\arguments{
\item{biopax}{The BioPaX object containing the parsed data from an OWL file. This can be obtained using readBiopax(filepath)}

\item{pathwayID}{The ID of the pathway in the biopax object}
}
\value{
A mully graph built from the given pathway
}
\description{
Build a mully graph from a given pathway
}
\note{
This should be preceded by readBiopax(filepath) to obtain the biopax object
}
\examples{
biopax=readBiopax("pi3k.owl")
pi3kmully=pathway2mully(biopax,"pathway1")
}
