% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Views.R
\name{pathwayView}
\alias{pathwayView}
\title{Create an empty view}
\usage{
pathwayView(g, name)
}
\arguments{
\item{g}{The input graph}

\item{name}{The name of the view}
}
\value{
An empty view
}
\description{
Create an empty view
}
\examples{
\dontrun{ 
view=pathwayView(mully("myMully",T),"View1")
}
}
