% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Views.R
\name{print.pathwayView}
\alias{print.pathwayView}
\title{Print function}
\usage{
\method{print}{pathwayView}(pathwayView)
}
\arguments{
\item{pathwayView}{The input View to be printed}
}
\value{

}
\description{
Print function
}
