% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pseudobulks.R
\name{remove_duplicate_rows}
\alias{remove_duplicate_rows}
\title{Remove Duplicate Rows}
\usage{
remove_duplicate_rows(mat, rownames_dups, method = c("max", "sum", "mean"))
}
\arguments{
\item{mat}{a matrix or sparse matrix of numeric data, such as a gene expression matrix.}

\item{rownames_dups}{a vector of rownames for `mat` that contains duplicates.}

\item{method}{which approach to use to aggregate data across duplicated rows (see:Details).}
}
\value{
a matrix with the provided rownames but without duplicates.
}
\description{
Removes rows from a matrix where there are duplicate row names, e.g. when changing gene IDs between databases or species.
}
\details{
Combines or selects one row of a matrix to represent all rows with duplicated rownames. There are three approaches:
max : keep the row with the highest average across columns
sum : add up the rows with duplicated rownames for each column.
mean: take the average across the duplicated rows for each column
}
\examples{
example_data <- generate_test_cellcounts()$counts
gene_ids <- sample(paste("gene", 1:10, sep=""), 20, replace=TRUE)
deduped <- remove_duplicate_rows(example_data, gene_ids, method="max")
total <- remove_duplicate_rows(example_data, gene_ids, method="sum")
avg <- remove_duplicate_rows(example_data, gene_ids, method="mean")
dim(deduped) == dim(total) # TRUE
dim(deduped) == dim(avg) # TRUE
dim(deduped)[1] == length(unique(gene_ids)) # TRUE
}
