% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hierarchical_Pathways.R
\name{select_hierarchy_term_pairs}
\alias{select_hierarchy_term_pairs}
\title{Hierarchical Pathways}
\usage{
select_hierarchy_term_pairs(
  depth,
  n,
  N = 50,
  max_term_size = 1000,
  seed = 2938,
  genes = NULL,
  gene_anno,
  n_attempts = 100
)
}
\arguments{
\item{depth}{the number of steps through the ontology separating each pair of pathways.}

\item{n}{the number of GO term pairs to select}

\item{N}{minimum number of genes per pathway.}

\item{max_term_size}{the maximum genes per pathway to consider that pathway}

\item{seed}{the random seed to ensure reproducibility}

\item{genes}{a background set of genes to limit pathways to.}

\item{gene_anno}{table of gene->GOterm annotations (from: load_GOBP_info())}

\item{n_attempts}{the maximum number of attempts to make to find pathways that fit all the criteria.}
}
\value{
a vector of GO term IDs, pairs are sequential: 1,2,1,2,1,2,1,2,...
}
\description{
Select random pairs of pathways separated in the GO hierarchy by a specified number of steps.
}
\details{
Uses pre-calculated lists of terms at each depth in the GOBP ontology (using GO.db package for ontology structure) to randomly select 'n' pathways. Then walks up the ontology the specified 'depth' number of steps to select the second pathway. Various additional constraints can be imposed on the selected pathways. This function makes 'n_attempts' attempts to find pairs of pathways that fit all the constraints.
}
\examples{
gene2go <- load_GOBP_info()
tissue_specific <- get_tissue_specific_genes(1, seed=301)
terms <- select_hierarchy_term_pairs(depth=1, 2, N=10, max_term_size=1000, seed=2938, genes=tissue_specific, gene_anno=gene2go)
terms <- select_hierarchy_term_pairs(depth=5, 2, N=10, max_term_size=1000, seed=2938, genes=tissue_specific, gene_anno=gene2go)
child_terms = terms[seq(from=1, to=length(terms)-1, by=2)]
parent_terms = terms[seq(from=2, to=length(terms), by=2)]
}
