% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hierarchical_Pathways.R
\name{select_random_terms}
\alias{select_random_terms}
\title{Random Pathways}
\usage{
select_random_terms(
  n,
  N = 50,
  max_term_size = 1000,
  seed = 2938,
  genes = NULL,
  gene_anno
)
}
\arguments{
\item{n}{the number of GO terms to select}

\item{N}{minimum number of genes per pathway.}

\item{max_term_size}{the maximum genes per pathway to consider that pathway}

\item{seed}{the random seed to ensure reproducibility}

\item{genes}{a background set of genes to limit pathways to.}

\item{gene_anno}{table of gene->GOterm annotations (from: load_GOBP_info())}
}
\value{
a vector of GO term IDs fitting the requirements.
}
\description{
Select random tissue-specific pathways.
}
\details{
Using various contraints randomly samples valid GO pathways.
}
\examples{
gene2go <- load_GOBP_info()
tissue_specific <- get_tissue_specific_genes(1, seed=301)
terms <- select_random_terms(5, N=20, max_term_size=1000, seed=2938, genes=tissue_specific, gene_anno=gene2go)
}
