% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hierarchical_Pathways.R
\name{select_shared_pathways}
\alias{select_shared_pathways}
\title{Select Shared Pathways}
\usage{
select_shared_pathways(
  nc,
  ns,
  nu,
  N = 50,
  max_term_size = 1000,
  seed = 2938,
  genes = NULL,
  gene_anno
)
}
\arguments{
\item{nc}{the number of comparisons}

\item{ns}{the number of pathways shared amongst all comparisons}

\item{nu}{the number of pathways unique to a comparison}

\item{N}{minimum number of genes per pathway.}

\item{max_term_size}{the maximum genes per pathway to consider that pathway}

\item{seed}{the random seed to ensure reproducibility}

\item{genes}{a background set of genes to limit pathways to.}

\item{gene_anno}{table of gene->GOterm annotations (from: load_GOBP_info())}
}
\value{
a list of tables of pathways and whether or not the pathway is shared or unique.
}
\description{
Generates a set of shared and unique pathways for modelling a multiple comparison
}
\details{
Simulates a situation where multiple comparisons have been performed for one experiment and pathway analysis has been performed for each comparison. It assumes some pathways are shared amongst all comparisons and some are unique to each individual comparison. Unique pathways are drawn independently for each comparison thus may overlap but are always distinct from the shared pathways.
}
\examples{
gene2go <- load_GOBP_info()
tissue_specific <- get_tissue_specific_genes(1, seed=301)
term_list <- select_shared_pathways(3, 2, 5, N=20, max_term_size=1000, seed=2938, genes=tissue_specific, gene_anno=gene2go)
term_list[[1]] # pathways for comparison 1
shared <- term_list[[1]][ term_list[[1]][,2] == "shared" ,1]
unique1 <- term_list[[1]][ term_list[[1]][,2] == "unique" ,1]
}
