% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG.R
\name{simplifyInteractionKegg}
\alias{simplifyInteractionKegg}
\title{Simplify the dataframe's structure}
\usage{
simplifyInteractionKegg(genes)
}
\arguments{
\item{genes}{A data frame of all genes and their relation to other cross references}
}
\value{
A data frame of gene and its presence in other databases
}
\description{
Simplify the dataframe's structure
}
\note{
must be preceeded by interactionKegg(keggInput)
}
\examples{
geneList=c("hsa:122706","hsa:4221","hsa:8312")
genes = getKeggGene(geneList)
genesinteraction = interactionKegg(genes)
genesinteractionsimplified =simplifyInteractionKegg(genesinteraction)
}
\author{
Mohammad Al Maaz
}
