% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrichments.R
\docType{data}
\name{test_paths}
\alias{test_paths}
\title{Teat Pathways}
\format{
a list of 50 pathways:
\describe{
   \item{item name}{name of the pathway}
   \item{value}{vector of human gene IDs}
 }
}
\source{
{msigdb} Bioconductor database package containing MSigDB pathways.
}
\usage{
test_paths
}
\description{
MSigDb Hallmark pathways
}
\keyword{datasets}
