% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG.R
\name{transformKeggData}
\alias{transformKeggData}
\title{Transform data retrieved from KEGG}
\usage{
transformKeggData(list_keggGet, list_Get)
}
\arguments{
\item{list_keggGet}{list of gene input}

\item{list_Get}{list of gene data retrieved from Kegg genes database}
}
\value{
a dataframe of Kegg data
}
\description{
Transform data retrieved from KEGG
}
\note{
must be precedeed by getkegggenes(geneInput)
}
\examples{
geneList=c("hsa:122706","hsa:4221","hsa:8312")
genes=  getKeggGene(geneList)
}
