% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pseudobulks.R
\name{trim_for_pseudobulks}
\alias{trim_for_pseudobulks}
\title{Trim Pseudobulk Samples}
\usage{
trim_for_pseudobulks(clusters, individual, nmin = 10)
}
\arguments{
\item{clusters}{a vector of cell-type names or cluster IDs.}

\item{individual}{a vector of individual or sample IDs.}

\item{nmin}{the minimum number of cells required for a reliable pseudobulk estimate.}
}
\value{
a vector of cluster-individual pairs that should be excluded
}
\description{
Removes sample x cell-type combinations where there are too few cells to obtain reliable pseudobulk estimates.
}
\details{
Any combination of cell-type x individual (or sample) that has fewer than `nmin` cells are identified for removal. This is called by get_pseudobulks.
}
\examples{
example_data <- generate_test_cellcounts()
trim_for_pseudobulks(example_data$celltypes, example_data$donors) # all of celltype1 should be identified for removal.
}
