% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrichments.R
\name{trim_pathway_names}
\alias{trim_pathway_names}
\title{Trim Pathway Names}
\usage{
trim_pathway_names(out, prefix_length = 1, nwords = 4, split_char = "_")
}
\arguments{
\item{out}{a set of pathway enrichments for a single DE test obtained from do_ora or do_fgsea}

\item{prefix_length}{the number of 'words' that make up the prefix you want to remove.}

\item{nwords}{the maximum number of words in the final pathway name.}

\item{split_char}{the regular expression to use to split the pathway name into individual words}
}
\value{
pathway enrichments provided in the "out" argument with an additional column in the results that includes the trimmed pathway name.
}
\description{
Trims the length of pathway names.
}
\details{
Removes a prefix and truncates the pathway names to a specified maximum number of words. 
This improves plot appearance when very long pathways are selected.
}
\examples{
#paths <- convert_GSEAObj_to_list(get_pathways("test"))
paths <- test_paths; print("6")
rich <- do_ora(paths[[1]], paths, background=unique(unlist(paths)))
rich <- condense_terms(rich)
rich <- trim_pathway_names(rich)
}
