% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Views.R
\name{undo}
\alias{undo}
\title{Undo a modification step in a view}
\usage{
undo(v, stps = 1)
}
\arguments{
\item{v}{The input view}

\item{stps}{The number of steps to undo. This number referes to the number of unique steps' IDs to be removed, i.e. entries of steps in the view with similar stepID count as 1}
}
\value{
The view with the undone modifications
}
\description{
Undo a modification step in a view
}
