% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnt_pathway.R
\name{wntpathway}
\alias{wntpathway}
\title{Demo function for Wnt Pathway Views}
\usage{
wntpathway(file)
}
\arguments{
\item{file}{The link to the Wnt Pathway biopax file}
}
\description{
Demo function for Wnt Pathway Views
}
