% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG.R
\name{addGenesLayer}
\alias{addGenesLayer}
\title{Add KEGG Gene Layer with its respective Nodes}
\usage{
addGenesLayer(g, biopax)
}
\arguments{
\item{g}{An existing Mully graph}

\item{biopax}{A biopax file}
}
\value{
A Mully graph with the all layers,nodes and edges
}
\description{
Add KEGG Gene Layer with its respective Nodes
}
\examples{
biopax=readBiopax("wnt.owl") #This is a mully graph that contains a protein layer
pathwayID=listPathways(biopax)$id[1]
g=Multipath::pathway2Mully(biopax,pathwayID)
g=addGenesLayer(g,biopax)
}
\author{
Mohammad Al Maaz
}
