% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrichments.R
\name{condense_terms}
\alias{condense_terms}
\title{Condense Terms}
\usage{
condense_terms(
  out,
  equivalent = 0.5,
  verbose = FALSE,
  prioritize.signaling = TRUE
)
}
\arguments{
\item{out}{a set of pathway enrichments for a single DE test obtained from do_ora or do_fgsea}

\item{equivalent}{a scalar value for the equivalence threshold.}

\item{verbose}{boolean, whether to print data table used as the basis for selecting representative terms.}

\item{prioritize.signaling}{a boolean, indicating whether pathways recieve a bonus for being signalling pathways}
}
\value{
The same structured data like the output from do_ora or do_fgse but with synonmyous termscondensed into a single pathway
}
\description{
Removes synonymous pathways from a set of output pathway enrichments
}
\details{
First calculates overlaps using 'get_overlaps' then identifies groups of synonymous pathways using 'cluster_overlaps' then selects the best representative term for each group using 'select_term'.
}
\examples{
paths <- convert_GSEAObj_to_list(get_pathways("test"))
rich <- do_ora(paths[[1]], paths, background=unique(unlist(paths)))
rich <- condense_terms(rich)
}
