% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrichments.R
\name{get_overlaps}
\alias{get_overlaps}
\title{Pathway Overlaps}
\usage{
get_overlaps(out)
}
\arguments{
\item{out}{a set of pathway enrichments for a single DE test obtained from do_ora or do_fgsea}
}
\value{
A matrix of overlap scores for all pairs of pathways.
}
\description{
Calculates the degree of overlap between genes contributing to each pathway enrichment.
}
\details{
Overlaps are calculated as the length of the intersection divided by the total genes in the larger pathway.
}
\examples{
paths <- convert_GSEAObj_to_list(get_pathways("test"))
rich <- do_ora(paths[[1]], paths, background=unique(unlist(paths)))
overlaps <- get_overlaps(rich)
}
