% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG.R
\name{getRelatedGenes}
\alias{getRelatedGenes}
\title{Get proteins that has a reference to KEGG Genes from a mully graph and a biopax}
\usage{
getRelatedGenes(g, biopax)
}
\arguments{
\item{g}{A Mully graph}

\item{biopax}{A biopax file}
}
\value{
List of Protein
}
\description{
Get proteins that has a reference to KEGG Genes from a mully graph and a biopax
}
\examples{

proteinList = c("P02747","P00734","P07204","A0A0S2Z4R0","O15169")
geneList=c("hsa:122706","hsa:4221","hsa:8312")
relatedGenes = getRelatedGenes(g,biopax)
}
