% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniProtKB.R
\name{getUPKBInfo}
\alias{getUPKBInfo}
\title{Get Proteins from UniProtKB}
\usage{
getUPKBInfo(proteins, col)
}
\arguments{
\item{proteins}{The list of UniProtKB Proteins ID to be retrieved}

\item{col}{The list of attributes associated to the UniProtKB Entries to be retrieved}
}
\value{
a dataframe containing the protein entries with the selected attributes
}
\description{
Get Proteins from UniProtKB
}
\note{
To get the list of possible columns, call queryup::returnfields and check the field column"
}
\examples{
\dontrun{ 
getUPKBInfo(c("Q6ZS62","P14384"),c("protein_name","go")) 
} 
}
