% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multipath.R
\name{multipath}
\alias{multipath}
\title{Generate Multipath Graph from General Data}
\usage{
multipath(
  name = "Multipath",
  up = NA,
  proteinList = NA,
  data = NA,
  drugList = NA
)
}
\arguments{
\item{name}{The name of the graph to be generated}

\item{up}{The Uniprot.ws() object}

\item{proteinList}{}

\item{data}{}

\item{drugList}{}
}
\value{
A mully graph with the added data
}
\description{
Generate Multipath Graph from General Data
}
