% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multipath.R
\name{multipath}
\alias{multipath}
\title{Generate Multipath Graph from General Data}
\usage{
multipath(name = "Multipath", proteinList = NA, data = NA, drugList = NA)
}
\arguments{
\item{name}{The name of the graph to be generated}

\item{proteinList}{The list of proteins of which the interactions should be retrieved}

\item{data}{The dataframe containing the parsed information of DrugBank. This argument can be obtained using the function loadDBXML(DrugBankFile)}

\item{drugList}{The list of DrugBank Ids of the drugs. This argument can be either a string (one drug) or a list of strings (multiple drugs)}
}
\value{
A mully graph with the added data
}
\description{
Generate Multipath Graph from General Data
}
