% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Views.R
\name{print.pathwayView}
\alias{print.pathwayView}
\title{Print function}
\usage{
\method{print}{pathwayView}(x, ...)
}
\arguments{
\item{x}{The input View to be printed}

\item{...}{any other parameteres passed to \code{print}}
}
\description{
Print function
}
