% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrichments.R
\name{select_term}
\alias{select_term}
\title{Select Term}
\usage{
select_term(out, terms, verbose = FALSE, path_scores = NULL)
}
\arguments{
\item{out}{a set of pathway enrichments for a single DE test obtained from do_ora or do_fgsea}

\item{terms}{a vector of names of pathways that are synonymous}

\item{verbose}{boolean, whether to print data table used as the basis for selecting the term.}

\item{path_scores}{a dataframe containing the columns: "pathway", "Score", "FDR". This is used primarily used as part of condensed_terms_multi, to prioritize pathways first by the Score, then in the case of tied Scores by FDR.}
}
\value{
The name of one pathway to represent the group.
}
\description{
Selects one pathway to represent a group of pathways
}
\details{
Given a set of synonymous pathways, this function selects one pathway to represent the group. Each pathway is scored based on the length of the pathway name, the p-value of the pathway enrichment, the size of the intersection between the pathway and the DE gens, and whether or not the pathway is a signalling pathway. Each value is ranked across pathways using the "min" method to resolve ties.
These ranks are summed and the best scoring pathway is selected.
}
\examples{
#paths <- convert_GSEAObj_to_list(get_pathways("test"))
paths <- test_paths;print("3")
rich <- do_ora(paths[[1]], paths, background=unique(unlist(paths)));
overlaps <- get_overlaps(rich);
pathway_groups <- cluster_overlaps(overlaps)
chosen1 <- select_term(rich, rich$results[pathway_groups == "1","pathway"])
chosen2 <- select_term(rich, rich$results[,"pathway"])
}
