% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnt_pathway.R
\name{wntpathway}
\alias{wntpathway}
\title{Demo function for Wnt Pathway Views}
\usage{
wntpathway()
}
\description{
Demo function for Wnt Pathway Views
}
\note{
This function calls "data", which is supposed to be the parsed DrugBank object. Please refer to loadDBXML().
}
